refactor(client): 알림 행·아이콘·이동 로직을 notifShared 모듈로 추출
NotifRow·typeVisuals·formatAgo 와 chat 링크 호환 이동 로직이 NotificationBell
안에만 있어 다른 화면(곧 추가할 전용 알림 페이지)에서 재사용하려면 복붙해야 했다.

- components/notifShared.tsx 신설: NotifRow, navigateToNotif 공개 / typeVisuals,
  formatAgo 는 모듈 내부 구현 디테일로 유지
- handleClick 의 인라인 chat 링크 분기를 navigateToNotif(n, nav) 호출로 대체

동작 변화 없음(순수 리팩터). 빌드 통과.
